+++ CHANGELOG.md
@@ -19,6 +19,11 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 - [#378](https://github.com/CarbonLDP/carbonldp-js-sdk/issues/378) - Allow pagination and ordering of multiple `$get()` documents
 	- The pagination sub-select (that may repeat the entire query) will now only be added when pagination (`limit()`, `offset()`) and/or sorting (`orderBy()`) are used.
 	  This also applies to the `$getChildren()` and `$getMembers()` methods
+	  
+### Fixed
+
+- [#391](https://github.com/CarbonLDP/carbonldp-js-sdk/issues/391) - Adding properties with `undefined`, `null` and/or `[]` (empty array) triggers a patch request on save, even though this doesn't represent RDF changes
+- [#392](https://github.com/CarbonLDP/carbonldp-js-sdk/issues/392) - Assigning a property that only wraps its previous value into an array triggers a patch request on save, even though this doesn't represent RDF changes
 
 ## [5.2.0] - 2019-04-10